@@ -3,10 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^15.6.1",
+    "react": "19.1.1",
     "react-audio-player": "^0.6.1",
-    "react-dom": "^15.6.1",
-    "react-responsive-modal": "^1.7.1"
+    "react-dom": "19.1.1",
+    "react-responsive-modal": "^7.0.0"
   },
   "scripts": {
     "start": "vite",
